Changed character set to UTF-8

+++ RImageBook/build.sh
@@ -1,12 +1,17 @@
 #!/bin/sh
-
-./chcharset.sh -w || exit 1
-mv NAMESPACE NAMESPACE.bak
-cp NAMESPACE.linux NAMESPACE
-pushd ..; R CMD build RImageBook ; popd
-mv NAMESPACE.bak NAMESPACE
-./chcharset.sh -s || exit 1
-pushd ..
-echo "install.packages(\"RImageBook_1.0.tar.gz\")" | R --no-save
-popd
+case $OS in
+Windows*)
+  echo This script is for linux OS.
+  ;;
+*)
+  mv NAMESPACE NAMESPACE.bak || exit 1
+  cp NAMESPACE.linux NAMESPACE || exit 1
+  pushd ..
+  if [ ! -e RImageBook_1.0.tar.gz ]; then
+    R CMD build RImageBook || exit 1
+  fi
+  R CMD install RImageBook_1.0.tar.gz
+  popd 
+  mv NAMESPACE.bak NAMESPACE
+esac
